requirements.txt
# CloudGuard-B Phase 1 — Requirements
# =============================================

# ── Core Framework ──
fastapi>=0.110.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# ── Data & Math ──
numpy>=1.26.0,<3.0.0
networkx>=3.2.0,<4.0.0

# ── Infrastructure (optional — graceful fallback if missing) ──
redis>=5.0.0,<6.0.0
celery>=5.4.0,<6.0.0
# asyncpg>=0.30.0,<1.0.0      # Uncomment for PostgreSQL async
# psycopg2-binary>=2.9.0,<3.0.0  # Uncomment for PostgreSQL sync

# ── Original CloudGuard Dependencies ──
elasticsearch>=8.12.0,<9.0.0
boto3>=1.34.0,<2.0.0
faker>=23.2.0,<30.0.0
requests>=2.31.0,<3.0.0

# ── Phase 2 Brain (optional — graceful fallback if missing) ──
# chromadb>=0.4.0,<1.0.0           # H-MEM vector store
# langgraph>=0.0.30                # State machine orchestrator
# langchain>=0.1.0,<1.0.0          # LangChain primitives
# langchain-core>=0.1.0,<1.0.0     # Core abstractions
google-generativeai>=0.4.0       # Gemini 1.5 Pro API
httpx>=0.27.0,<1.0.0               # Ollama HTTP client + testing

# ── Testing ──
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
